requirements.txt
PyVirtualDisplay==1.3.2
PyYAML==5.3.1
selenium==3.141.0
urllib3==1.25.11
